RecruitmentPipeline_Package/requirements.txt
requests
PyYAML
pypdf
python-docx
openai
argparse
python-dotenv
pyinstaller